临时提交

+++ src/aspnet-core/FlyFramework.Common/Helpers/JWTTokens/JwtBearerModel.cs
@@ -20,6 +20,6 @@
         /// <summary>
         /// 过期时间（分钟）
         /// </summary>
-        public double AccessTokenExpiresMinutes { get; set; }
+        public int AccessTokenExpiresMinutes { get; set; }
     }
 }
